Allow tests to be run with pytest
