+++ tests/test_work_managers/test_serial.py
@@ -1,7 +1,9 @@
+import unittest
+
 from westpa.work_managers.serial import SerialWorkManager
 from .tsupport import CommonWorkManagerTests
 
 
-class TestSerialWorkManager(CommonWorkManagerTests):
+class TestSerialWorkManager(unittest.TestCase, CommonWorkManagerTests):
     def setUp(self):
         self.work_manager = SerialWorkManager()
